requirements.txt
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
bleach==4.0.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
decorator==5.0.9
defusedxml==0.7.1
dicttoxml==1.7.4
diff-match-patch==20200713
Django==3.2.6
django-import-export==2.5.0
django-jazzmin==2.4.8
django-user-agents==0.4.0
docutils==0.17.1
et-xmlfile==1.1.0
future==0.18.2
geocoder==1.38.1
geoip2==2.9.0
idna==3.2
importlib-metadata==4.6.4
ip2geotools==0.1.5
IP2Location==8.6.2
isort==5.9.3
keyring==23.1.0
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupPy==1.14
maxminddb==2.0.3
mccabe==0.6.1
memcache==0.8.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
Pillow==8.3.1
pip-review==1.1.0
pkginfo==1.7.1
pycodestyle==2.7.0
Pygments==2.10.0
pylightxl==1.56
pylint==2.9.6
pyparsing==2.4.7
pyquery==1.4.3
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
ratelim==0.1.6
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
tqdm==4.62.1
twine==3.4.2
typed-ast==1.4.3
typing==3.7.4.3
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
